requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
boto3
python-multipart
python-jose[cryptography]
passlib[bcrypt]